Rearranged the way the form works.
It works the following way.
The VM has some asynchronous work to do to figure out if there's an existing value for the reading or a new one.

A mutable state value for the reading hold the data as it's supposed to be.
This will be set either asynchronously to the value to be edited or if that fails, to an empty object.

Once the object is not empty anymore, the UI for the editor will load.

Here the values are stored in remember saveables so that they can persist via destructon of state.

Once all editing is done, the composeable creates a new value from them and sends it on for saving.